Added --enable-extension-checks configuration option
Compiled with this option, Ruby will warn when a `T_DATA` object is
created whose class has not undefined or redefined the `allocate`
method.

A new function is defined, `rb_data_object_check`. That function is
called from `rb_data_object_wrap()` and
`rb_data_typed_object_wrap()` (which implement the `Data_Wrap_Struct`
family of macros).

This option is intended to help developers of C extensions to check if
their code meets the requirements explained in "doc/extension.rdoc".

Without a check like this, there is no easy way for an author of a C
extension to easily see where they have made a mistake.