Warn unless T_DATA object classes redefine or undefine `allocate`
Ruby will warn when a `T_DATA` object is created whose class has not
undefined or redefined the `allocate` method.

This option is intended to help developers of C extensions to check if
their code meets the requirements explained in "doc/extension.rdoc".

Without a check like this, there is no easy way for an author of a C
extension to easily see where they have made a mistake.